L2WormholeGateway: addressed Slither warnings
Two warnings in receiveWormhole were false-positives and got addressed
by adding slither-disable-next-line.

The function is non-reentrant and we need to call
completeTransferWithPayload first to obtain the amount and payload.

The warning in sendWormhole was also a false-positive given the function
is non-reentrant and the L2TBTC is a trusted contract. I was able to
address this warning by changing the order of calls instead of ignoring
it, though.