Removed the wallet member authorization in WalletCoordinator
Having every wallet member be able to submit sweep proposals to the wallet
enables the risk of griefing the coordination protocol. The address authorized
to submit proposals can keep submitting proposals that wallet members will
reject off-chain. Given that we do not perform the validation on-chain during
the proposal submission to optimize gas coordination costs, every address
eligible to submit a proposal can block the wallet for a certain timeout
specific to the proposed action. Addresses appointed by the DAO as coordinators
can be stripped of the coordinator role if they abuse the protocol. It is not
possible to remove the address from the wallet member list though. A possible
solution for the problem of a malicious wallet member griefing the coordination
is introducing a rotating role of a wallet member coordinator, as proposed in
RFC 7: Sweeping coordination. Given that we are not going to use this mechanism
in the first iteration of sweeping and redemptions and given that the
WalletCoordinator contract is supposed to be deployed behind an upgradeable
transparent proxy, the ACL allowing wallet members to submit proposals has been
removed. If we decide to enable wallet members to submit proposals in the
future, we will probably do it as specified in RFC 7: Sweeping coordination, by
introducing a rotating wallet member coordination role.