Changing a backoff retrier for the system tests
withBackoffRetrier implementation will try to call a bitcoin client
again in case of the thrown error. A lot of times it happens that there
is no error in response, but rather no data is returned. This commit
hanlde this case as well and fixes the system tests, which were failing
for this reason.