Update dependencies to `random-beacon`, `ecdsa` and `tbtc-v2`
We have published new `@keep-network/random-beacon`, `@keep-network/ecdsa` and
`@keep-network/tbtc-v2` packages. Those packages no longer include the
`prepare-dependencies.sh` script from `@threshold-network/solidity-contracts`
which was causing random failures during `yarn install`. As in some CI jobs we
install the dependencies based on the lockfile (with the `--frozen-lockfile`
flag), we need to update dependencies in the lockfile so that the new, improved
packages would be used.