Update the `TBTCToken` interface
Add the `approveAndCall` function that calls `receiveApproval` function on
spender previously approving the spender to withdraw from the caller
multiple times, up to the `amount` amount. If this function is called
again, it overwrites the current allowance with `amount`. We need this
function because we want to call redemption request in one tx instead of
two(approve + requestRedemption).