Update dependencies to `tbtc-v2` and `tbtc-v2.ts` packages
We've published new versions of the `@keep-network/tbtc-v2` and
`@keep-network/tbtc-v2.ts` packages The versions of tghose packages for the
development environment don't contain `prepare-dependencies.sh` script from
`@threshold-network/solidity-contracts` which was causing random failures
during `yarn install`. The `goerli` and `mainnet` packages still contain that
script, so we won't completely eliminate the problem, but should decrease its
frequency.
In a separate commit we'll add a workaround that should help with eliminating
the problem in CI.