Run postinstall script separately from `yarn install`
Sometimes during installation of the `monitoring` project we were getting random
failures related to the presence of the `prepare-dependencies.sh` script from
`threshold-network/solidity-contracts` in the project subdependencies. Lately we
managed to remove the script from the `development`-tagged packages, but
the script is still present in `goerli` and `mainnet` packages. As `monitoring`
project installs `goerli` and `mainnet` `@keep-network/tbtc-v2` packages during
`yarn install`, the random failures may still occur. As a workaround for this
problem in our CI we will run install step with the `--ignore-scripts` flag
which installs the dependencies without running the postinstall scripts (neither
in subdependencies nor in the main project). We then add a step that executes
`yarn run postinstall` - this way we still execute any postinstall scripts from
the main project (currently it's `npm rebuild bcrypto`) script.