Deployment script for WalletCoordinator upgrade
We add a script for WalletCoordinator contract upgrade.

The current WalletCoordinator contract deployed to mainnet is owned by
Threshold Council, so it cannot be upgraded right away by the deployer.

The deployer will deploy a fresh WalletCoordinator contract
implementation, then Threshold Council need to execute transactions to
upgrade the WalletCoordinator proxy to the new implementation and set
the initial values for newely added parameters.