Prototyping canonical SolanaTBTC token with the Solang compiler
Solang compiles Solidity code that can be run on Solana blockchain.
SolanaTBTC contract uses the spl_token lib which is ERC20 equivalent on
Solana. There were a couple of functionalities added to match L2TBTC.sol
contract that was deployed to Arbitrum, Optimims and Polygon. However,
the contracts that are on L2s do not match 1:1 with this one. E.g. there
is no 'recoverERC20' function. Upgrade and pause (freeze) functionality
should work out of the box on Solana, hence there are no functions
written for it.

setup.ts was largely copied from Solang official repository.