Fix system tests
Request the redemption via the `depositorBridgeHandle.requestRedemption`
directly instead of  `requestRedemption` function. After refactoring in
`#632`, the `requestRedemption` requests redemption via tBTC token
contract. In the current deposit scenario, we do not pass the vault
address so the `depositor` does not actually have any tBTC tokens so we
can't request redemption via `requestRedemption` fn with changes from
`#632`. We are going to add a new scenario where we pass the TBTC vault
address to the `revealDeposit` function to test the new mechanism of
requesting the redemption via tBTC token contract in follow-up work.