Import Signer from `ethers` lib
In the past there was an issue with recognizing the `Signer` in the tbtc-v2.ts
lib. As a workaround I've used `Signer` from `@ethersproject/abstract-signer`.
The issue was fixed in `25d2de2ab89ee3c8428d7044f4f80fe53b15f750` so we can now
safely import SIgner from `ethers`.