Wrap wallet-api calls with try-catch block
As you know before any wallet-api call we first have to connect() our
`windowsMessageTransport` and `disconnect()` it right after to avoid any
potential problems. However, there might be a situation when something goes
wrong with the call and the next line, which is transport disconnect, does not
execute, which may leas to some issues later on.

This is why we additionaly wrap those calls calls try and cattch block, where we
disconnect the `windowMessageTransport` if an error occures.