Add Sepolia support for Polygon cross-chain
The Goerli testnet currently used by Threshold/Keep for development purposes is
planned to become deprecated with the end of year 2023. The testnet that was
created to replace it is called
[Holešky](https://github.com/eth-clients/holesky), however it will take some
time until it gets integrated with by some of the projects we rely on.
As a solution, we decided to switch first to another testnet that is currently
live - Sepolia. This testnet's EOL is planned for 2026, which gives us plenty of
time to move to Holešky before Sepolia gets deprecated.
Until Görli is not dead we want to support both testnets.

The Goerli -> Sepolia migration means that also L2 testnet basing on Goerli
(Mumbai) needs to be migrated to Sepolia-based chain. At the moment the new
testnet isn't yet announced, so deployment of our contract is yet not possible.
But we're already starting to add changes that prepare us for the moment when it
will be supported.
Once new testnet is live and WOrmhole support is confirmed, we'll need to verify
if `wormholeChainID` we use in the config is correct for the testnet. We'll also
need to update the missing Hardhat configuration (TODOs) and add the
`export-artifacts:sepolia` spript to `package.json`.