Introduce support for depositor proxies
Here we introduce the concept of a `DepositorProxy` contract. A `DepositProxy`
can reveal deposits to the Bridge, on behalf of the user
(i.e. original depositor). Once minting is completed, the `DepositProxy`
receives minted TBTC and can provide additional services to the user, such as
routing the minted TBTC tokens to another protocols, in an automated way.

To expose this feature to the SDK users, we are adding a new
`initiateDepositWithProxy` function. This function triggers the deposit flow
with support of the given `DepositProxy`.