Make `...WormholeChainID` values dependable on the environment
Wormhole has added support for Arbitrum Sepolia, Base Sepolia and Optimism
Sepolia testnets. We want to modify the config of our deployment scripts so that
deployment on Sepolia-based L2s would be possible. As the Chain IDs for Wormhole
Sepolia-based networks differ from from IDs for Mainnet and Goerli-based
testnets, we need to set the right value of `...WormholeChainID` constant for
the the right network we're deploying to.