Do not emit events from AbstractTBTCDepositor
In all cases considered so far, the child contract emits its own events in the
functions calling `_initializeDeposit` and `_finalizeDeposit`. The events
emitted by child contracts have the advantage of emitting decoded extra data in
the parameters. We could emit events from the parent contract as well but they
cost gas and we are emitting the same data in child contract events. Also, we
sometimes need to be creative in child contracts and come up with other names
than DepositInitialized and DepositFinalized even though that is exactly what
happens :)