Turn on the watchtower and manage guardians
Here we implement functions allowing to enable the watchtower (i.e. veto
mechanism) and manage guardians.

The `enableWatchtower` function allows setting the watchtower manager and an
initial set of guardians. Moreover, we are capturing the initiation timestamp
which will be useful while determining pre-OR and post-OR redemption requests
and will be needed to implement the planned lifecycle of the veto mechanism

Regarding guardian management, we are adding two functions:
- `addGuardian` that can be called by the watchtower manager (Token Holder DAO)
- `removeGuardian` that can be called by the owner (Threshold Council)