Retrun initial deposit amount from _initializeDeposit function
The amount may be useful for integrators for validations of the deposit
amount. See example use case in https://github.com/thesis/acre/pull/253#discussion_r1500680454

It turns out that this version where we read from the storage costs
`117 366` gas. Which is less than an alternative approach:
```
        initialDepositAmount =
            fundingTx
                .outputVector
                .extractOutputAtIndex(reveal.fundingOutputIndex)
                .extractValue() *
            SATOSHI_MULTIPLIER;
```
which costs `117 601` gas.