Draft implementation of the `L1BitcoinDepositor` contract
Here we present a draft implementation of the `L1BitcoinDepositor` contract
that acts as the central point of the tBTC direct bridging feature on the
L1 Ethereum chain where TBTC minting occurs. This contract exposes the
`receiveWormholeMessages` function that receives a message from
`L2BitcoinDepositor` contract (using the Wormhole Relayer infrastructure),
extracts deposit data from it, and initiates the deposit on the tBTC `Bridge`
side. Moreover, the contract also exposes the `initializeDeposit` function
directly. The goal here is to satisfy future use cases that don't rely on
Wormhole Relayer for cross-chain messaging.