Change `calldata` to `memory` in `AbstractTBTCDepositor` contract
So far, the `_initializeDeposit` accepted `calldata` parameters. This
does not work for integrator contracts that obtain deposit parameters
as `memory` objects (the conversion `memory -> calldata` is not possible).

To overcome this problem, we are changing parameters of `_initializeDeposit`
to be `memory` as well. This is not breaking integrators that receive
deposit parameters as `calldata` because the `calldata -> memory` conversion
is fine.