Introduce redemption with proxy
This commit introduces a new redemption flow that uses a proxy implementation
to request the redemption of tBTC. It gives flexibility to the external
integrators to use the redemption data required by the tBTC protocol to
request the redemption of tBTC in their contracts.
It can also be used to relay the transaction submission with external
relayer (e.g. Gelato).